BUGFIX - byte handling error

+++ app/src/main/java/com/gruutnetworks/gruutsigner/ui/dashboard/DashboardViewModel.java
@@ -351,9 +351,9 @@ public class DashboardViewModel extends AndroidViewModel implements LifecycleObs
         String signature;
         try {
             // TODO check this out later. format 맞추기
-            byte[] sigSender = ByteBuffer.allocate(8).putInt(Integer.parseInt(sender)).array();
-            byte[] sigTime = ByteBuffer.allocate(8).putInt(Integer.parseInt(time)).array();
-            byte[] sigHgt = ByteBuffer.allocate(8).putInt(Integer.parseInt(msgRequestSignature.getBlockHeight())).array();
+            byte[] sigSender = ByteBuffer.allocate(8).putLong(Integer.parseInt(sender)).array();
+            byte[] sigTime = ByteBuffer.allocate(8).putLong(Integer.parseInt(time)).array();
+            byte[] sigHgt = ByteBuffer.allocate(8).putLong(Integer.parseInt(msgRequestSignature.getBlockHeight())).array();
 
             ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
             outputStream.write(sigSender);
